+++ src/babel/__tests__/examples/zacs.styled(builtin).js
@@ -1,5 +1,9 @@
 import * as zacs from '@nozbe/zacs'
 
-// you wouldn't do it in shared code, only in .{native,web}.js
+// web builtins
 const StyledLi = zacs.styled('li')
 const li = <StyledLi foo={foo} title="hello" />
+
+// react-native builtins
+const StyledRCTText = zacs.styled('RCTText')
+const text = <StyledRCTText foo={foo} title="hello" />